requirements.txt
fastapi[all]
sqlalchemy
python-jose
passlib
bcrypt